verifying 64-bit allocations, #367, done for algos, compositing, core, filters, gui and io directories

+++ src/algos/statistics_float.c
@@ -115,7 +115,7 @@ static double siril_stats_float_bwmv(const float* data, const size_t n,
 	return bwmv;
 }
 
-int IKSS(float *data, long n, double *location, double *scale, gboolean multithread) {
+int IKSS(float *data, size_t n, double *location, double *scale, gboolean multithread) {
 	size_t i, j;
 	double mad, s, s0, m;
 	float xlow, xhigh;
